feat: 对外暴露一些表格内部的取数方法
