+++ src/biz/utils/exportTableAsExcel.ts
@@ -1,8 +1,8 @@
 import XLSX_NS from 'xlsx'
-import { ArtColumn } from '../../interfaces'
-import collectNodes from '../../common-utils/collectNodes'
-import { getTreeDepth, safeGetValue } from '../../common-utils/internals'
+import { collectNodes, getTreeDepth } from '../../common-utils'
+import { safeGetValue } from '../../common-utils/internals'
 import isLeafNode from '../../common-utils/isLeafNode'
+import { ArtColumn } from '../../interfaces'
 
 type XlsxCellDatum = string | number | null
