+++ src/common-utils/index.tsx
@@ -1,3 +1,8 @@
 export { default as collectNodes } from './collectNodes'
+export { default as getTreeDepth } from './getTreeDepth'
 export { default as groupBy } from './groupBy'
 export { default as isLeafNode } from './isLeafNode'
+
+import * as internals from './internals'
+
+export { internals }
