+++ src/pivot/cross-table/buildCrossTable.tsx
@@ -1,8 +1,7 @@
 import cx from 'classnames'
 import React from 'react'
 import { BaseTableProps } from '../../base-table'
-import { isLeafNode } from '../../common-utils'
-import { getTreeDepth } from '../../common-utils/internals'
+import { getTreeDepth, isLeafNode } from '../../common-utils'
 import { ArtColumn, SpanRect } from '../../interfaces'
 import { CrossTableClasses, ROW_KEY } from './constants'
 import { CrossTableProps } from './cross-table'
